Extract modules and classes to separate files
I found it a bit overwhelming and confusing to have all
functionality (scrapers, models and helper classes) in a
single file.

Here I've made separate files for each of them, for easier
overview, hopefully.
