+++ lib/hpr/additional_expertise.rb
@@ -1,0 +1,3 @@
+module Hpr
+  class AdditionalExpertise < Struct.new(:name, :period); end
+end
