+++ lib/hpr/specials.rb
@@ -1,0 +1,5 @@
+require "hpr/additional_expertise"
+
+module Hpr
+  class Specials < AdditionalExpertise; end
+end
